--- src/main/java/swyp12/team9/server/domain/jwt/dto/JwtResponse.java
@@ -1,12 +1,0 @@
-package swyp12.team9.server.domain.jwt.dto;
-
-import io.swagger.v3.oas.annotations.media.Schema;
-
-@Schema(description = "JWT 토큰 응답 DTO")
-public record JwtResponse(
-        @Schema(description = "Access 토큰", example = "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...")
-        String accessToken,
-        @Schema(description = "Refresh 토큰", example = "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...")
-        String refreshToken
-) {
-}
