+++ src/main/java/swyp12/team9/server/domain/reference/repository/ReferenceRepositoryImpl.java
@@ -13,7 +13,7 @@ import swyp12.team9.server.domain.reference.dto.response.ReferenceListResponse;
 import java.util.List;
 
 import static swyp12.team9.server.domain.reference.model.QReference.reference;
-import static swyp12.team9.server.domain.referenceuserlink.model.QReferenceUserLink.referenceUserLink;
+import static swyp12.team9.server.domain.reference.relation.model.QReferenceUserLink.referenceUserLink;
 import static swyp12.team9.server.domain.userlink.model.QUserLink.userLink;
 
 @RequiredArgsConstructor
