+++ src/main/java/swyp12/team9/server/domain/stat/service/UserStatService.java
@@ -5,8 +5,8 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import swyp12.team9.server.domain.stat.dto.*;
-import swyp12.team9.server.domain.referenceuserlink.dto.ReferenceCategoryCountProjection;
-import swyp12.team9.server.domain.referenceuserlink.repository.ReferenceUserLinkRepository;
+import swyp12.team9.server.domain.reference.relation.dto.ReferenceCategoryCountProjection;
+import swyp12.team9.server.domain.reference.relation.repository.ReferenceUserLinkRepository;
 import swyp12.team9.server.domain.userlink.dto.DayCountProjection;
 import swyp12.team9.server.domain.userlink.model.LinkStatus;
 import swyp12.team9.server.domain.userlink.repository.UserLinkRepository;
