+++ src/main/java/swyp12/team9/server/domain/userlink/service/UserLinkSearchService.java
@@ -9,8 +9,8 @@ import swyp12.team9.server.domain.userlink.dto.response.UserLinkSearchResponse;
 import swyp12.team9.server.domain.reference.exception.ReferenceNotFoundException;
 import swyp12.team9.server.domain.reference.model.Reference;
 import swyp12.team9.server.domain.reference.repository.ReferenceRepository;
-import swyp12.team9.server.domain.referenceuserlink.model.ReferenceUserLink;
-import swyp12.team9.server.domain.referenceuserlink.repository.ReferenceUserLinkRepository;
+import swyp12.team9.server.domain.reference.relation.model.ReferenceUserLink;
+import swyp12.team9.server.domain.reference.relation.repository.ReferenceUserLinkRepository;
 import swyp12.team9.server.domain.userlink.model.UserLink;
 import swyp12.team9.server.domain.userlink.repository.UserLinkSearchRepository;
