+++ src/main/java/swyp12/team9/server/global/handler/LoginSuccessHandler.java
@@ -9,7 +9,7 @@ import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.web.authentication.AuthenticationSuccessHandler;
 import org.springframework.stereotype.Component;
-import swyp12.team9.server.domain.jwt.service.JwtService;
+import swyp12.team9.server.domain.auth.jwt.service.JwtService;
 import swyp12.team9.server.domain.user.model.User;
 import swyp12.team9.server.domain.user.repository.UserRepository;
 import swyp12.team9.server.global.util.JwtUtil;
