+++ src/test/java/swyp12/team9/server/domain/user/service/UserStatServiceTest.java
@@ -8,8 +8,8 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import swyp12.team9.server.domain.stat.dto.UserStatResponse;
-import swyp12.team9.server.domain.referenceuserlink.dto.ReferenceCategoryCountProjection;
-import swyp12.team9.server.domain.referenceuserlink.repository.ReferenceUserLinkRepository;
+import swyp12.team9.server.domain.reference.relation.dto.ReferenceCategoryCountProjection;
+import swyp12.team9.server.domain.reference.relation.repository.ReferenceUserLinkRepository;
 import swyp12.team9.server.domain.stat.service.UserStatService;
 import swyp12.team9.server.domain.userlink.dto.DayCountProjection;
 import swyp12.team9.server.domain.userlink.model.LinkStatus;
